Added Test ipynb for better testing
