+++ backend/requirements.txt
@@ -4,3 +4,4 @@ flask-cors
 pandas
 xlrd
 openpyxl
+ipykernel
